feat(agents): surface running agents inside the active assistant message
Move the live agent-activity pill out of the chat status bar and into the
assistant message that's actually running the work. The status bar is for
chat-wide settings; per-turn activity belongs next to the response.

- New `AgentActivity` component renders only when there's active work
  (running / waiting / error). At rest the surface is fully invisible.
- `App.tsx` resolves the right host message (`agentActivityMessageID`)
  and passes the status snapshot to that `MessageView` only.
- `StatusBar` drops `AgentsMenu`, the `agents` popover ID, and the
  `agentsStatus` prop.
- Ported the existing AgentsMenu tests onto the new component; no
  popover-semantics regressions (running pill, error state, popover
  open/close, Escape, outside-click, grouping, subagent labels).

Closes #170
